Implement balance management service and transaction type enum; refactor transaction service for improved transaction handling and validation; update transaction input models and validators.

+++ API/Services/Transaction/TransactionService.cs
@@ -1,3 +1,4 @@
+using API.Domain.BalanceManagement;
 using API.Extensions;
 using API.Inputs;
 using API.Models;
@@ -6,27 +7,46 @@ using API.Utils.UserContext;
 using Mapster;
 using Microsoft.EntityFrameworkCore;
 
-namespace API.Services;
+namespace API.Services.Transaction;
 
-public class TransactionService(AppDbContext context, ICurrentUserContext currentUserContext, NotificationContext notificationContext) : ITransactionService
+public class TransactionService(AppDbContext context, ICurrentUserContext currentUserContext, NotificationContext notificationContext, IBalanceManagementService balanceManagementService) : ITransactionService
 {
-    public async Task<TransactionDto?> CreateTransaction(CreateTransactionInput transactionCreateDto)
+    public async Task<TransactionDto?> CreateTransaction(CreateTransactionInput createTransactionInput)
     {
         var userId = currentUserContext.RequiredUserId;
 
-        var transaction = transactionCreateDto.Adapt<Models.Transaction>();
+        await using var dbTransaction = await context.Database.BeginTransactionAsync();
+
+        var transaction = createTransactionInput.Adapt<Models.Transaction>();
         transaction.UserId = userId;
 
         var addedTransaction = context.Transactions.Add(transaction);
 
-        await context.SaveChangesAsync();
+        if (addedTransaction == null)
+        {
+            notificationContext.AddNotification("Failed to create transaction", ErrorType.ServerError);
+            return null;
+        }
 
-        var createdTransaction = await context.Transactions
+        var source = await context.Sources.FirstOrDefaultAsync(s => s.Id == addedTransaction.Entity.SourceId && s.UserId == userId);
+
+        if (source == null)
+        {
+            notificationContext.AddNotification("Source not found", ErrorType.NotFound);
+            return null;
+        }
+
+        if (!await balanceManagementService.UpdateAmounts(addedTransaction.Entity, source, userId))
+            return null;
+
+        await context.SaveChangesAsync();
+        await dbTransaction.CommitAsync();
+
+        return await context.Transactions
         .AsNoTracking()
         .ProjectToType<TransactionDto>()
-        .FirstOrDefaultAsync(t => t.Id == addedTransaction.Entity.Id && t.UserId == userId);
+        .FirstOrDefaultAsync(t => t.Id == transaction.Id && t.UserId == userId);
 
-        return createdTransaction;
     }
 
     public async Task<bool> DeleteTransaction(int id)
@@ -40,8 +60,71 @@ public class TransactionService(AppDbContext context, ICurrentUserContext curren
             return false;
         }
 
+        var dbTransaction = await context.Database.BeginTransactionAsync();
+
+        if (transaction.TransactionTypeId == (int)TransactionTypeEnum.Adjustment)
+        {
+            notificationContext.AddNotification("Adjustment transactions cannot be deleted", ErrorType.BadRequest);
+            return false;
+        }
+
+        if (!await balanceManagementService.ResetTransaction(transaction, userId))
+            return false;
+
         context.Transactions.Remove(transaction);
         return await context.SaveChangesAsync() > 0;
+    }
+
+    public async Task<TransactionDto?> UpdateTransaction(int id, UpdateTransactionInput transactionUpdateDto)
+    {
+        var userId = currentUserContext.RequiredUserId;
+
+        var transaction = context.Transactions.FirstOrDefault(t => t.Id == id && t.UserId == userId);
+
+        var originalTransaction = transaction.Adapt<Models.Transaction>();
+
+        if (transaction == null)
+        {
+            notificationContext.AddNotification("Transaction not found", ErrorType.NotFound);
+            return null;
+        }
+
+        if (transaction.TransactionTypeId == (int)TransactionTypeEnum.Adjustment)
+        {
+            notificationContext.AddNotification("Adjustment transactions cannot be updated", ErrorType.BadRequest);
+            return null;
+        }
+
+        await using var dbTransaction = await context.Database.BeginTransactionAsync();
+        var updatedTransaction = transactionUpdateDto.AdaptIgnoreNull(transaction);
+
+        if (updatedTransaction.Amount != originalTransaction.Amount || updatedTransaction.TransactionTypeId != originalTransaction.TransactionTypeId || updatedTransaction.SourceId != originalTransaction.SourceId || updatedTransaction.DestinationSourceId != originalTransaction.DestinationSourceId)
+        {
+            if (!await balanceManagementService.ResetTransaction(originalTransaction, userId))
+                return null;
+
+            var source = await context.Sources.FirstOrDefaultAsync(s => s.Id == updatedTransaction.SourceId && s.UserId == userId);
+            if (source == null)
+            {
+                notificationContext.AddNotification("Source not found", ErrorType.NotFound);
+                return null;
+            }
+
+            if (!await balanceManagementService.UpdateAmounts(updatedTransaction, source, userId))
+                return null;
+
+        }
+
+        await context.SaveChangesAsync();
+        await dbTransaction.CommitAsync();
+
+        var updatedTransactionDto = await context.Transactions
+        .AsNoTracking()
+        .ProjectToType<TransactionDto>()
+        .FirstOrDefaultAsync(t => t.Id == id && t.UserId == userId);
+
+        return updatedTransactionDto;
+
     }
 
     public async Task<TransactionDto?> GetTransactionById(int id)
@@ -69,9 +152,10 @@ public class TransactionService(AppDbContext context, ICurrentUserContext curren
         var transactions = context.Transactions
         .AsNoTracking()
         .ProjectToType<TransactionDto>()
-        .Where(t => t.UserId == userId)
-        .Skip(input.Offset)
-        .Take(input.Limit);
+        .Where(t => t.UserId == userId);
+
+        if (input.CategoryId != null)
+            transactions = transactions.Where(t => t.Category != null && t.Category.Id == input.CategoryId);
 
         if (input.FromDate != null)
             transactions = transactions.Where(t => t.Date >= input.FromDate);
@@ -92,30 +176,9 @@ public class TransactionService(AppDbContext context, ICurrentUserContext curren
             };
         }
 
-        return await transactions.ToListAsync(); ;
-    }
-
-    public async Task<TransactionDto?> UpdateTransaction(int id, UpdateTransactionInput transactionUpdateDto)
-    {
-        var userId = currentUserContext.RequiredUserId;
-
-        var transaction = context.Transactions.FirstOrDefault(t => t.Id == id && t.UserId == userId);
-
-        if (transaction == null)
-        {
-            notificationContext.AddNotification("Transaction not found", ErrorType.NotFound);
-            return null;
-        }
-
-        transactionUpdateDto.AdaptIgnoreNull(transaction);
-
-        await context.SaveChangesAsync();
-
-        var updatedTransaction = await context.Transactions
-        .AsNoTracking()
-        .ProjectToType<TransactionDto>()
-        .FirstOrDefaultAsync(t => t.Id == id && t.UserId == userId);
-
-        return updatedTransaction;
+        return await transactions
+        .Skip(input.Offset)
+        .Take(input.Limit)
+        .ToListAsync();
     }
 }
